refactor: extract VDF parser into standalone crate
