+++ crates/vdf-parser/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "vdf-parser"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.79"
+authors = ["Travis Lane"]
+description = "A parser for Steam's VDF file formats including appinfo.vdf and shortcuts.vdf"
+license = "GPL-3.0"
+repository = "https://github.com/Tormak9970/Steam-Art-Manager"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+rayon = "1.10.0"
