+++ crates/vdf-parser/src/vdf_structs.rs
