--- src-tauri/src/parsers/mod.rs
@@ -1,4 +1,0 @@
-mod vdf_structs;
-mod vdf_reader;
-pub mod appinfo_vdf_parser;
-pub mod shortcuts_vdf_parser;
