+++ src-tauri/src/utils/mod.rs
@@ -1,3 +1,1 @@
 pub mod logger;
-pub mod reader;
-pub mod writer;
